+++ apps/app/src/electron/EverythingService.ts
