+++ apps/app/src/electron/triggersHandler.ts
@@ -4,7 +4,7 @@ import _ from 'lodash'
 import { getGroupPlayData } from '../lib/playout/groupPlayData'
 import { ActiveTrigger, ActiveTriggers } from '../models/rundown/Trigger'
 import { BridgeHandler } from './bridgeHandler'
-import { EverythingService } from './IPCServer'
+import { EverythingService } from './EverythingService'
 import { StorageHandler } from './storageHandler'
 import {
 	ActionAny,
